Buff Logic terminée et variation de Damage Mult et coup crit
=> passer sur la generation des pattern changeant selon le coup
    ==> baser le fonctionnement sur un systeme qui prend en compte le pattern complet en commençant selement avec la tile pattern[0]

=> pour la selection : le button manager est deja ovverride pour afficher le reseau
    ==> donc en profiter pour modif le UnitManager pôur ajouter un isNetworkOn et permettre de cliquer sur le reseau
    ==> pour les lanceurs, afficher simplement autour de la tile sélectionnée les GetNeighbours

+++ Assets/Scripts/UI/InGameUI/DamageFeedBack.cs
@@ -19,7 +19,7 @@ namespace UI.InGameUI
 
         public void OnInit(float damage)
         {
-            damageText.text = "-" + damage;
+            damageText.text = "-" + (int)damage;
             damageText.color = Color.red;
             transform.DOMoveY(transform.position.y + 0.75f, 2);
         }
